backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
email-validator
pydantic[email]
pytest
httpx
